pyproject.toml
[tool.poetry]
name = "test_task_01"
version = "0.1.0"
description = ""
authors = ["ilderebezov <ilderebezov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.2"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
